Improve TS fix and move SocialIcon SVGs
